// ui/src/api/generated/model/serverProcessInfo.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */

export interface ServerProcessInfo {
  pid: number;
  name: string;
  username?: string | null;
  cpu_percent: number;
  memory_percent: number;
  status: string;
  create_time?: string | null;
}